rename game modules
