+++ js/get-game-tripple.js
