+++ js/render-random-game.js
@@ -1,19 +1,19 @@
-import getGame1 from './module_game-1';
-import getGame2 from './module_game-2';
-import getGame3 from './module_game-3';
+import getGameDouble from './get-game-double';
+import getGameSingle from './get-game-single';
+import getGameTripple from './get-game-tripple';
 
 const getRandomNumber = (min, max) => Math.floor(min + Math.random() * (max + 1 - min));
 const renderRandomGame = () => {
   const gameModule = getRandomNumber(1, 3);
   switch (gameModule) {
     case 1:
-      getGame1();
+      getGameSingle();
       break;
     case 2:
-      getGame2();
+      getGameDouble();
       break;
     case 3:
-      getGame3();
+      getGameTripple();
       break;
   }
 };
